UI revision
* UI revision
- Removed all sliders. (HP, ATK, DEF)
- Added STATPOINT
- Added new texts for HP, ATK, DEF
- More varaibles passing in UpdatePlayerStatus

+++ Assets/Scripts/Gameplay/Resources/EnemyData/BaseTemplate/Enemy_AttackDetail.cs
@@ -7,7 +7,7 @@ public class Enemy_AttackDetail
 {
     public string EnemyAttackName;
     public bool PhysicalAttack;
-    public int EnemyAttackDmg;
+    public float EnemyAttackDmg;
     public int EnemyAttackType;
 
 }
